Add debian/ dir & build packages in CI (for Debian and derivatives)
Keep in mind that:
- when you update/set the version for a release, it has to be updated in debian/changelog as well.
- when you update the flutter version you should also update the version in debian/rules. The build will fail by design if it is not the same version as the one in your github actions job.

For now I have disabled the tests because one test fails. It can be reenabled once the tests PASS.

Note that this cannot stricly follow the "Debian policy" since (among other reasons), to build the binary package, we have to download flutter SDK from the internet as well as rustup for older distributions.

The github actions job produces the binary packages (*.deb) for 6 suites:
- Debian 11 bullseye
- Debian 12 bookworm
- Debian 13 trixie
- Ubuntu 22.04 jammy
- Ubuntu 24.04 noble
- Kali rolling

I can confirm that it runs perfectly on Debian bookworm. I can't test the other, but they should work as well.